+++ internal/github/github_client.go
@@ -1,0 +1,19 @@
+package github
+
+import (
+	"context"
+	"os"
+
+	"github.com/google/go-github/v32/github"
+	"golang.org/x/oauth2"
+)
+
+func NewGitHubClient() *github.Client {
+	ctx := context.Background()
+	ts := oauth2.StaticTokenSource(
+		&oauth2.Token{AccessToken: os.Getenv("GHORG_GITHUB_TOKEN")},
+	)
+	tc := oauth2.NewClient(ctx, ts)
+	client := github.NewClient(tc)
+	return client
+}